new app orders created and cart model added to it

+++ myfab/main/models.py
@@ -46,11 +46,12 @@ class Product(models.Model):
     price = models.DecimalField(max_digits=10, decimal_places=2, default=0, validators=[MinValueValidator(0)])
     image = models.ImageField(upload_to='images/products')
     description = models.TextField(blank=True)
-    qty = models.FloatField(default=0, validators=[MinValueValidator(0)])
+    qty = models.DecimalField(max_digits=10, decimal_places=2, default=0, validators=[MinValueValidator(0)])
     add_date = models.DateField(auto_now_add=True)
     is_active = models.BooleanField(default=True)
     usages = models.ManyToManyField(Usage)
     category_choices = models.ManyToManyField(CategoryChoice)
+
 
     def __str__(self):
         return self.name
